pipeline-manager: refactor and clean up
Removes:
- Retry function
- Metrics endpoint
- Unused dependencies

Moves:
- Demo data structure
- Database probe
- Database notifier
