+++ Cargo.toml
@@ -61,7 +61,6 @@ async-trait = "0.1"
 atomic = "0.6.0"
 awc = "3.5.1"
 aws-config = "1.1.7"
-aws-sdk-cognitoidentityprovider = "1.81.0"
 aws-sdk-s3 = "1.86.0"
 aws-types = "1.1.7"
 base64 = "0.22.1"
