+++ crates/pipeline-manager/src/db/mod.rs
@@ -1,7 +1,9 @@
 pub mod error;
+pub mod notifier;
 pub mod operations;
 #[cfg(feature = "postgresql_embedded")]
 mod pg_setup;
+pub mod probe;
 pub(crate) mod storage;
 pub mod storage_postgres;
 #[cfg(test)]
